Rakudo architecture
NQP is genuinely a distinct language in the compiler toolchain. It's essentially a boostrapping language for implementing other programming languages. It's close to the VM but high level enough to easily write parsers declaratively. The Rakudo compiler has identical optimisations for any target language for example. It optimises the Perl 6 semantics, NQP is then more like an abstract language near to each VM that implements what Perl 6 needs with language specific code for each low level operation. NQP is self hosting in its implementing languages Rakudo is not.